requirements.txt
google-play-scraper==1.2.7
app-store-web-scraper==0.2.0
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.3
openpyxl==3.1.5
requests==2.32.3
python-dotenv==1.0.1
functions-framework==3.8.2
google-cloud-storage==2.18.2
